FISH-242 Working test with multiple modules
