--- appserver/tests/payara-samples/samples/microprofile-healthcheck/ear/pom.xml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>fish.payara.samples.microprofile</groupId>
-        <artifactId>healthcheck</artifactId>
-        <version>5.23.0-SNAPSHOT</version>
-    </parent>
-
-    <groupId>fish.payara.samples.microprofile.healthcheck</groupId>
-    <artifactId>healthcheck-ear</artifactId>
-    <packaging>ear</packaging>
-
-    <name>Healthcheck Ear</name>
-  
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-ear-plugin</artifactId>
-                <version>3.0.2</version>
-                <configuration>
-                    <version>7</version>
-                    <applicationName>Healthcheck EAR</applicationName>
-                    <displayName>Healthcheck Ear</displayName>
-                    <generateApplicationXml>false</generateApplicationXml>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-    <dependencies>
-        <dependency>
-            <groupId>fish.payara.samples.microprofile.healthcheck</groupId>
-            <artifactId>healthcheck-ejb</artifactId>
-            <version>${project.version}</version>
-            <type>ejb</type>
-        </dependency>
-        <dependency>
-            <groupId>fish.payara.samples.microprofile.healthcheck</groupId>
-            <artifactId>healthcheck-web1</artifactId>
-            <version>5.23.0-SNAPSHOT</version>
-            <type>war</type>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>fish.payara.samples.microprofile.healthcheck</groupId>
-            <artifactId>healthcheck-web2</artifactId>
-            <version>${project.version}</version>
-            <type>war</type>
-        </dependency>
-        <dependency>
-            <groupId>fish.payara.samples</groupId>
-            <artifactId>test-utils</artifactId>
-        </dependency>
-        
-    </dependencies>
-</project>
