--- appserver/tests/payara-samples/samples/microprofile-healthcheck/ejb/pom.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>fish.payara.samples.microprofile</groupId>
-        <artifactId>healthcheck</artifactId>
-        <version>5.23.0-SNAPSHOT</version>
-    </parent>
-
-    <groupId>fish.payara.samples.microprofile.healthcheck</groupId>
-    <artifactId>healthcheck-ejb</artifactId>
-    <version>5.23.0-SNAPSHOT</version>
-    <packaging>ejb</packaging>
-    
-    <name>Healthcheck EJB</name>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.eclipse.microprofile.health</groupId>
-            <artifactId>microprofile-health-api</artifactId>
-            <version>2.2</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>javax</groupId>
-            <artifactId>javaee-api</artifactId>
-            <version>7.0</version>
-            <scope>provided</scope>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-ejb-plugin</artifactId>
-                <version>3.1.0</version>
-                <configuration>
-                    <ejbVersion>3.1</ejbVersion>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-    
-</project>
