+++ appserver/tests/payara-samples/samples/microprofile-healthcheck/pom.xml
@@ -10,19 +10,23 @@
     <groupId>fish.payara.samples.microprofile</groupId>
     <artifactId>healthcheck</artifactId>
     <name>Healthcheck example</name>
-    <packaging>pom</packaging>
     
-    <modules>
-        <module>ejb</module>
-        <module>web1</module>
-        <module>web2</module>
-        <module>ear</module>
-    </modules>
     <dependencies>
         <dependency>
             <groupId>org.jboss.shrinkwrap</groupId>
             <artifactId>shrinkwrap-api</artifactId>
-            <version>1.2.6</version>
+        </dependency>
+        <dependency>
+            <groupId>jakarta.platform</groupId>
+            <artifactId>jakarta.jakartaee-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.eclipse.microprofile.health</groupId>
+            <artifactId>microprofile-health-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>fish.payara.samples</groupId>
+            <artifactId>test-utils</artifactId>
         </dependency>
     </dependencies>
 </project>
