--- appserver/tests/payara-samples/samples/microprofile-healthcheck/web2/src/main/webapp/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Start Page</title>
-        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    </head>
-    <body>
-        <h1>Hello World!</h1>
-    </body>
-</html>
