+++ appserver/tests/payara-samples/samples/pom.xml
@@ -16,29 +16,28 @@
 
     <modules>
         <!-- This module must run first, because it collides with some change made by other test(s) -->
-<!--        <module>payara-expression-config-properties</module>
-       <module>asadmin</module>
+        <module>payara-expression-config-properties</module>
+        <module>asadmin</module>
         <module>clustered-singleton</module>
         <module>custom-loginmodule-realm</module>
         <module>datagrid-encryption</module>
         <module>dynamic-roles</module>
         <module>ejb-invoker-secure-endpoint</module>
         <module>formauth</module>
-
         <module>jacc-per-app</module>
         <module>jaxrs-rolesallowed</module>
         <module>jaxrs-rolesallowed-servlet</module>
-        <module>jaxws-security</module> 
+        <module>jaxws-security</module>
         <module>jaxws-tracing</module>
-        <module>microprofile-endpoints</module> -->
+        <module>microprofile-endpoints</module>
         <module>microprofile-healthcheck</module>
         <module>oauth2</module>
-        <!-- <module>openid</module>
+        <module>openid</module>
         <module>rest-management</module>
         <module>rolesPermitted</module>
         <module>realm-identity-stores</module>
         <module>jakartaee-namespace-transformer</module>
-        <module>logging</module> -->
+        <module>logging</module>
     </modules>
 
     <dependencyManagement>
